enhance installation conditions

+++ install.php
@@ -1,10 +1,12 @@
 <?php
 /* Tablesets aktualisieren */
 
-rex_yform_manager_table_api::importTablesets(rex_file::get(rex_path::addon($this->name, 'install/rex_event.tableset.json')));
+if (rex_addon::get('yform')->isAvailable() && !rex::isSafeMode()) {
+    rex_yform_manager_table_api::importTablesets(rex_file::get(rex_path::addon($this->name, 'install/rex_event.tableset.json')));
+    rex_yform_manager_table::deleteCache();
+}
 
-rex_yform_manager_table::deleteCache();
-
+/*
 $modules = scandir(rex_path::addon("dummy")."module");
 
 foreach ($modules as $module) {
@@ -24,3 +26,4 @@ foreach ($modules as $module) {
     ->setValue("updatedate", date("Y-m-d H:i:s"))
     ->insertOrUpdate();
 }
+*/
